apps/services/llm/external/openai/realtime/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
python-dotenv==1.0.0
jinja2==3.1.2
aiohttp==3.8.6
websockets>=10.0,<12.0
asyncio==3.4.3
httpx==0.25.0
python-multipart==0.0.6
pydantic==2.5.2
typing-extensions==4.8.0
